requirements.txt
# requirements.txt

torch
torchvision
tqdm
Pillow
requests
PyYAML
matplotlib
